refactor(R2.2): 搜索结果定位的纯逻辑摘出 ReadBookViewModel 并补测
章内搜索域和前三个域形态不同，按实际边界分两半处理，本提交只做能干净
摘出的那一半。ReadBookViewModel 5018 → 4906 行。

- searchResultPositions / findSearchResultMatch / findSearchTextPoint
  共 115 行是纯函数（只吃 TextChapter + SearchResult + query，不碰任何
  UI 状态），摘成 ReadSearchResultLocator。VM 保留 3 行转发入口，
  ReadBookController 的调用点零改动。file-private 的 SearchTextPoint
  随之搬走。
- 补 ReadSearchResultLocatorTest（8 条）。这段逻辑原来埋在 VM 里没法测：
  它先信 SearchResult.queryIndexInChapter 直达下标，校验不过才退回
  「数到第 N 次出现」的扫描。直达下标来自搜索时那一版正文，被替换规则或
  AI 净化改写后就失效——校验分支要是坏了，搜索跳转会静默跳到错的位置。
  已注入回归反证（把 if (directMatch) 改成 if (true)）确认会变红。

未做（有意留下）：搜索的**状态**没有摘成 delegate。isShowingSearchResult
/ searchResultList / searchResultIndex 被 ReadBookMenuBar、
ReadBookSearchBar、ReadBookController 返回键、ReadBookScreen 返回键、
MainNavGraph 五处直读——它是阅读器 chrome 的状态，不是弹层自包含状态，
和前三个域的边界不可比。强行摘会把 4 个 state 参数穿进 chrome 的重组热
路径，收益为负。留待 R2.3 Track B 端游时连同 chrome 状态一起处理。

验证：compileAppDebugKotlin + verifyConfigArchitecture + 全量单测
(426 passed) + assembleAppDebug + lintAppDebug 全绿。
